+++ async_v20/definitions/types.py
@@ -732,13 +732,18 @@ class TradeOpen(Model):
             The number of units opened by the Trade
         client_extensions: :class:`~async_v20.ClientExtensions`
             The client extensions for the newly opened Trade
+        initial_margin_required: :class:`~async_v20.AccountUnits`
+            The margin required at the time the Trade was created. Note, this is the
+            ‘pure’ margin required, it is not the ‘effective’ margin used that
+            factors in the trade risk if a GSLO is attached to the trade.
 
     """
 
     def __init__(self, price: DecimalNumber = sentinel, trade_id: TradeID = sentinel, units: DecimalNumber = sentinel,
                  client_extensions: ClientExtensions = sentinel,
                  guaranteed_execution_fee: AccountUnits = sentinel,
-                 half_spread_cost: AccountUnits = sentinel):
+                 half_spread_cost: AccountUnits = sentinel,
+                 initial_margin_required: AccountUnits = sentinel):
         Model.__init__(**locals())
 
 
@@ -1338,7 +1343,8 @@ class AccountChangesState(Model):
                  margin_closeout_position_value: DecimalNumber = sentinel, withdrawal_limit: AccountUnits = sentinel,
                  margin_call_margin_used: AccountUnits = sentinel, margin_call_percent: DecimalNumber = sentinel,
                  orders: ArrayDynamicOrderState = sentinel, trades: ArrayCalculatedTradeState = sentinel,
-                 positions: ArrayCalculatedPositionState = sentinel):
+                 positions: ArrayCalculatedPositionState = sentinel,
+                 balance: AccountUnits = sentinel):
         Model.__init__(**locals())
 
 
@@ -2385,6 +2391,10 @@ class Trade(Model):
         initial_units: :class:`~async_v20.DecimalNumber`
             The initial size of the Trade. Negative values indicate
             a short Trade, and positive values indicate a long Trade.
+        initial_margin_required: :class:`~async_v20.AccountUnits`
+            The margin required at the time the Trade was created. Note, this is the
+            ‘pure’ margin required, it is not the ‘effective’ margin used that
+            factors in the trade risk if a GSLO is attached to the trade.
         current_units: :class:`~async_v20.DecimalNumber`
             The number of units currently open for the Trade. This
             value is reduced to 0.0 as the Trade is closed.
@@ -2420,6 +2430,7 @@ class Trade(Model):
 
     def __init__(self, id: TradeID = sentinel, instrument: InstrumentName = sentinel, price: PriceValue = sentinel,
                  open_time: DateTime = sentinel, state: TradeState = sentinel, initial_units: DecimalNumber = sentinel,
+                 initial_margin_required: AccountUnits = sentinel,
                  current_units: DecimalNumber = sentinel, realized_pl: AccountUnits = sentinel,
                  unrealized_pl: AccountUnits = sentinel, average_close_price: PriceValue = sentinel,
                  closing_transaction_ids: ArrayTransactionID = sentinel, financing: AccountUnits = sentinel,