+++ CHANGELOG.rst
@@ -9,8 +9,8 @@ Changelog
  - Added guaranteed to StopLossOrder definition
  - Replaced ellipsis (...) with sentinel object
  - Added fullVWAP to Transaction definition
- - UnknownValue Exception is now a Warning in order to make the client more robust
- to breaking changes from OANDA
+ - UnknownValue Exception is now replaced with UnknownKeywordArgument Warning
+ in order to make the client more robust to breaking changes from OANDA
  - Checked many class definitions against updated OANDA documentation
 
 7.0.1b0